pyproject.toml
[tool.poetry]
name = "conference-room-reservation"
version = "0.1.0"
description = ""
authors = ["edo <takaaki.edo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flake8 = "^7.0.0"
black = "^24.4.0"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
pymysql = "^1.1.0"
sqlalchemy = "^2.0.29"
streamlit = "^1.33.0"
requests = "^2.31.0"
python-multipart = "^0.0.9"
bcrypt = "^4.1.2"
python-jose = "^3.3.0"
passlib = "^1.7.4"
cryptography = "^42.0.5"
stripe = "^9.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 88